third_party/blink/web_tests/http/tests/security/mixedContent/insecure-fetch-on-dedicated-worker-allowed.https-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/insecure-fetch-in-dedicated-worker.js' was loaded over HTTPS, but requested an insecure resource 'http://example.test:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*'. This content should also be served over HTTPS.
This is a testharness.js-based test.
Harness: the test ran to completion.
